requirements.txt
line-bot-sdk==3.17.1
flask==3.0.0
pygsheets==2.0.6
firebase-admin==6.5.0